third_party/WebKit/LayoutTests/fast/js/dot-node-base-exception-expected.txt

This test checks that a correct exception is raised when calculating the base value of a dot expression fails.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (document.appendChild()).foobar() threw exception NotFoundError: An attempt was made to reference a Node in a context where it does not exist..
PASS successfullyParsed is true

TEST COMPLETE
